Implement ReadableStreamGenericReader mixin
