+++ reference-implementation/lib/helpers/miscellaneous.js
@@ -11,3 +11,14 @@ exports.rethrowAssertionErrorRejection = e => {
     }, 0);
   }
 };
+
+exports.mixin = (target, source) => {
+  const keys = Reflect.ownKeys(source);
+  for (let i = 0; i < keys.length; ++i) {
+    if (keys[i] in target) {
+      continue;
+    }
+
+    Object.defineProperty(target, keys[i], Object.getOwnPropertyDescriptor(source, keys[i]));
+  }
+};
